First the eye tracker is physically set up for reading, so thats good.
Deleted some unnecessary classes (aoi, aoi repos, linkfinder), changed
the eye tracker factory to return a TobiiTracker instance, and updated
the Gaze and Link classes to reflect the changes in the UML diagram.

I addded code to write IGazeResponses to a new Gaze and add it to the
Gaze repos, although the handler is not firing using the dummy tracker.

I had the plugin launching, but after changing the factory to return a
TobiiTracker instance, the plugin is not working. There is an error
about no "TobiiTracker" in java.library.path, but no matter what I try I
can't get it to work as a plugin. I can however get it to work using the
main in TobiiTacker in the development eclipse instance.