Got a successful link between managed c# handle and unmanaged c++
You need to use gcroot to declare a managed handle on an unmanaged type.
You need this for any unmanaged type, whether it's being dllexported or
not. gcroot is used when you want to hold on to a managed type inside a
native class. You can't just declare a manged type directly in a native
type using the hat ^ symbol. You must use a gcroot. This is a "garbage
collected root". So, while the gcroot (a native object) lives, the
garbage collector cannot collect this object. When the gcroot is
destroyed, it lets go of the reference, and the garbage collector is
free to collect the object (assuming it has no other references).